Move app::RenderTile from app/ui/editor -> app/render
